requirements.txt
opencv-python-headless
qreader
pillow
requests
numpy
streamlit
libgl1